Small Fix with About Update
Fixed the hover text to just display the county name and the value (in this case species abundance) with a small space between the two. Fixed the about page to show some basic example text with bold and italics formats. Colors are now to scale for the bird map and the all species map displays but it currently shows bird data. This is due to the non-generic programming on the bird coloring function. It should be made generic to display with any dataset selected. Our current limitation is that this coloring function uses colorNumeric() which (as its name suggests) only works with numeric data. A solution will be needed later as it becomes more generic.