Add user-dataprep files and Fixed MODZCTA
The first API designed to access MODZCTA accessed a seperate source of data with no geometries, lat, lon, or other mappable points. This was fixed by identifying the  query error (a task in itself) and finding the right fix from NYC Open Data. This is now solved and contains geometries. I also added the user-dataprep_template_[x] files by copying the template and tacking on the data set name in place of the [x].  Data is accessed through Git links as csv for saving for reference if reorganizing later.