Dataset 2 Functions, Dropdown List Not Yet Functional
Data associated with app2.R is functional. We can confirm that the user needs some very specific, very clean data prior to using it in the map_server function. our_responsibility.R was hardly touched, I think the structure of the map_server function and map_ui currently work well together and did not want to ruin it. What would be beneficial (I think) is combining the lists of polygons and points to produce a list of all polygon and point layers that could be turned on and off. My current thought is a sophisticated list of lists. Sophisticated here meaning, not too deep that map_server can't access the data needed but also deep enough to have distinct layers appear. This is just a thought, although small scale tests have begun.