Roads_Push
We have roads! This was accomplished by taking a random sample of a larger dataframe with all primary, secondary, and tertiary roads in NY state from the tigris package. See the tigris_imports.R for details. I would recommend loading only the _ny_sample_roads from the tigris_imports.R script if troubleshooting, otherwise, it takes a long time for the roads to load or display. Screenshots/snips were taken of the tabs at different points. The 3 main tabs are working perfectly with the edits to our for loop from Jason. I attempted to see what it would look like if combined the data into a list of polygons, polylines, and points since, this seems to be one of Jason's main goals. However, I was not successful. Errors can be recreated (if needed) for future testing in app.R and app2.R. In both cases, the "all" tab can be easily removed from the ui and server. It may benefit us to remove this during any presentations as well.